feat: acquire tunnel from worker (#155)
