+++ linkup-cli/src/commands/start.rs
@@ -39,8 +39,7 @@ pub async fn start(
     env_logger::init();
 
     let mut state = if fresh_state {
-        let is_paid = services::CloudflareTunnel::use_paid_tunnels();
-        let state = load_and_save_state(config_arg, args.no_tunnel, is_paid)?;
+        let state = load_and_save_state(config_arg, args.no_tunnel, true)?;
         set_linkup_env(state.clone())?;
 
         state
@@ -264,6 +263,7 @@ fn set_linkup_env(state: LocalState) -> Result<(), CliError> {
     Ok(())
 }
 
+// TODO: Remove this `is_paid` arg
 fn load_and_save_state(
     config_arg: &Option<String>,
     no_tunnel: bool,
